+++ packages/language-service/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @effect-dynamodb/language-service
 
+## 2.0.0
+
 ## 1.0.0
 
 ### Minor Changes